requirements.txt
cffi==1.17.1
cryptography==45.0.6
mysql-connector-python==9.6.0
pycparser==2.22
python-dotenv==1.2.1
